# Use standard merge algorithm when merging Values