While investigating a runtime/GC stall, @aclements noticed that `Values.Merge` was using a sub-optimal algorithm and that a traditional merge algorithm would likely perform better in general.  He also discovered that the benchmark we were using had a bug which caused the current implementation to appear better when it was actually quite worse in all but one case.

This switches the `Values.Merge` to use a standard merge algorithm.

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkValues_Merge-8            729849        47429         -93.50%
BenchmarkValues_MergeSame-8        22544         51528         +128.57%
BenchmarkValues_MergeSimilar-8     247516        49819         -79.87%

benchmark                          old allocs     new allocs     delta
BenchmarkValues_Merge-8            1              1              +0.00%
BenchmarkValues_MergeSame-8        0              1              +Inf%
BenchmarkValues_MergeSimilar-8     1              1              +0.00%

benchmark                          old bytes     new bytes     delta
BenchmarkValues_Merge-8            32768         32768         +0.00%
BenchmarkValues_MergeSame-8        0             32768         +Inf%
BenchmarkValues_MergeSimilar-8     32768         32768         +0.00%
```