@jwilder 2017-03-14
@e-dard Added some other benchmarks:

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkValues_Merge-8             725067        46914         -93.53%
BenchmarkValues_MergeDisjoint-8     23278         24094         +3.51%
BenchmarkValues_MergeSame-8         22181         51152         +130.61%
BenchmarkValues_MergeSimilar-8      250766        49441         -80.28%
BenchmarkValues_MergeUnevenA-8      5111          7141          +39.72%
BenchmarkValues_MergeUnevenB-8      7302          7265          -0.51%

benchmark                           old allocs     new allocs     delta
BenchmarkValues_Merge-8             1              1              +0.00%
BenchmarkValues_MergeDisjoint-8     1              1              +0.00%
BenchmarkValues_MergeSame-8         0              1              +Inf%
BenchmarkValues_MergeSimilar-8      1              1              +0.00%
BenchmarkValues_MergeUnevenA-8      0              1              +Inf%
BenchmarkValues_MergeUnevenB-8      1              1              +0.00%

benchmark                           old bytes     new bytes     delta
BenchmarkValues_Merge-8             32768         32768         +0.00%
BenchmarkValues_MergeDisjoint-8     32768         32768         +0.00%
BenchmarkValues_MergeSame-8         0             32768         +Inf%
BenchmarkValues_MergeSimilar-8      32768         32768         +0.00%
BenchmarkValues_MergeUnevenA-8      0             12288         +Inf%
BenchmarkValues_MergeUnevenB-8      12288         12288         +0.00%
```